Updated Report Draft Propositions and Constraints
Added directions of movement and top left and right planet checks to propositions.

Added constraints such that the rocket can't go up and down at the same time or right and left at the same time, as well as a check to see if the rocket needs to turn and go up.