Added More Detailed Comments. Fixed many Bugs
Removed completed and obsolete TODOs.

Removed fuel from rocket, as there is a universal fuel for both the binary values of fuel (arr[]) and if the rocket has fuel or not (fuel).

Added doc strings and sufficient comments.

Changed yellow print of Checkpoint to vibrant yellow, and regular yellow for deactivated Checkpoints.

Fixed rocket avoidance system.

Added SpaceObject in the way of the rocket in stage one to show avoidance system.

Encountered issue where b1-b8 in arr are not being set to True or False propositions, but rather set arr[i] to the value True or False. Not solved.