Added Reachability, Beacons, People have Unique Colour, Cleaned Code
Reachability constraint added. A cell is reachable if the cell is next to the rocket's path (left, right, down, up, diagonals included).

Added Beacons which will be used by the SAT solver to be placed in the most optimal way given a limited number of them (optimal meaning saves the most people, so the Beacon has a radius around it, and the people within that radius are "saved.")

People are now printed as True in light blue, and are no longer able to be saved from a planet they live on.

Co-Authored-By: Harsh Shrivastav <harsh.shrivastav1725@gmail.com>
Co-Authored-By: LeoTonerM <186430879+LeoTonerM@users.noreply.github.com>
Co-Authored-By: Nandanbhut <92518445+Nandanbhut@users.noreply.github.com>