User can no longer place SpaceObjects on other objects; Rocket Starts centered in Stage 1 and 2
User can no longer place SpaceObjects on people, planets, other SpaceObjects, or the starting position of the Rocket.

Rocket now starts at y = RADIUS - 1 for stages 1 and 3, as before it did not start at the center of the planet/grid as the comments had suggested.

Now prints initial grids with people before Rocket is added so the user can use them to decide where to place their SpaceObjects.