Removed Dead Fuel Code.
Fuel code is no longer being used as it does not increase the work needed to be done by the SAT Solver in its current state. At most the SAT Solver would have to find if fuel is True or False based on if any of the b1-b8 are True.